+++ Film_scraper/src/main/java/ru/spbstu/sql/GetHistory.sql
@@ -1,4 +1,4 @@
-create function get_history(curruserid integer) RETURNS setof integer
+create function get_history(curruserid bigint) RETURNS setof integer
     language sql
 as
 $$
@@ -6,4 +6,4 @@ SELECT film_id FROM schedule where id IN
         (SELECT schedule_id FROM history where user_id = currUserId)
 $$;
 
-alter function get_history(integer) owner to postgres;
+alter function get_history(bigint) owner to postgres;
